GC2D/SelectShine2 move: 93.86% -> 93.87% (keep frame limit arithmetic full width)

+++ src/GC2D/SelectShine2.cpp
@@ -450,7 +450,7 @@ void TSelectShine::move()
 		}
 	} else if (unk49 != 0) {
 		unk3C        = (s16)(unk3C + 8);
-		s16 maxFrame = (s16)(mAnmColor->mMaxFrame - 1);
+		s32 maxFrame = mAnmColor->mMaxFrame - 1;
 		if (unk3C > maxFrame) {
 			unk3C = maxFrame;
 		}
